fix m_misc.c(295): Error! E1010: Type mismatch (source conversion type is 'unsigned char *', target conversion type is 'int *')

+++ FASTDOOM/g_game.c
@@ -161,7 +161,7 @@ int dclicks2;
 int savegameslot;
 char savedescription[32];
 
-byte autorun;
+int autorun;
 
 //
 // G_BuildTiccmd
